LPD-51250 Separate space selector
