+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/categorization/tags/CreateTagsModal.tsx
@@ -4,24 +4,15 @@
  */
 
 import ClayButton from '@clayui/button';
-import {ClayCheckbox} from '@clayui/form';
-import ClayIcon from '@clayui/icon';
 import ClayModal from '@clayui/modal';
-import ClayMultiSelect from '@clayui/multi-select';
-import ClaySticker from '@clayui/sticker';
 import {useFormik} from 'formik';
 import {openToast} from 'frontend-js-components-web';
 import {fetch, navigate} from 'frontend-js-web';
-import React, {useEffect, useState} from 'react';
+import React from 'react';
 
 import {FieldText} from '../../components/forms';
 import {required, validate} from '../../components/forms/validations';
-import SpaceService from '../services/SpaceService';
-
-type Space = {
-	label: string;
-	value: any;
-};
+import CategorizationSpaces from '../components/CategorizationSpaces';
 
 export default function CreationTagModalContent({
 	assetLibraryId,
@@ -30,70 +21,51 @@ export default function CreationTagModalContent({
 	assetLibraryId?: string;
 	tagsURL: string;
 }) {
-	const [availableSpaces, setAvailableSpaces] = useState<Space[]>([]);
-	const [checkbox, setCheckbox] = useState(true);
-	const [allSpaces, setAllSpaces] = useState<Space[]>([]);
+	const {errors, handleChange, handleSubmit, resetForm, touched, values} =
+		useFormik({
+			initialValues: {
+				assetLibraryIds: [],
+				tagName: '',
+			},
+			onSubmit: (values) => {
+				const url =
+					'/o/headless-admin-taxonomy/v1.0/asset-libraries/' +
+					assetLibraryId +
+					'/keywords';
 
-	useEffect(() => {
-		SpaceService.getSpaces().then((response) => {
-			const spaces = response.map((space) => ({
-				label: space.name,
-				value: space.id,
-			}));
+				const body = {
+					name: values.tagName,
+				};
 
-			setAvailableSpaces(spaces);
-
-			setAllSpaces([
-				{
-					label: 'All Spaces',
-					value: response.map(({id}) => id),
-				},
-			]);
-		});
-	}, [assetLibraryId]);
-
-	const {
-		errors,
-		handleChange,
-		handleSubmit,
-		resetForm,
-		setFieldValue,
-		touched,
-		values,
-	} = useFormik({
-		initialValues: {
-			assetLibraryIds: [],
-			tagName: '',
-		},
-		onSubmit: (values) => {
-			const url =
-				'/o/headless-admin-taxonomy/v1.0/asset-libraries/' +
-				assetLibraryId +
-				'/keywords';
-
-			const body = {
-				name: values.tagName,
-			};
-
-			fetch(url, {
-				body: JSON.stringify(body),
-				headers: {
-					'Accept': 'application/json',
-					'Content-Type': 'application/json',
-				},
-				method: 'POST',
-			})
-				.then((response) => {
-					if (response.ok) {
-						openToast({
-							message: Liferay.Language.get(
-								'your-request-completed-successfully'
-							),
-							title: Liferay.Language.get('success'),
-							type: 'success',
-						});
-					}
-					else {
+				fetch(url, {
+					body: JSON.stringify(body),
+					headers: {
+						'Accept': 'application/json',
+						'Content-Type': 'application/json',
+					},
+					method: 'POST',
+				})
+					.then((response) => {
+						if (response.ok) {
+							openToast({
+								message: Liferay.Language.get(
+									'your-request-completed-successfully'
+								),
+								title: Liferay.Language.get('success'),
+								type: 'success',
+							});
+						}
+						else {
+							openToast({
+								message: Liferay.Language.get(
+									'an-unexpected-error-occurred'
+								),
+								title: Liferay.Language.get('error'),
+								type: 'danger',
+							});
+						}
+					})
+					.catch(() => {
 						openToast({
 							message: Liferay.Language.get(
 								'an-unexpected-error-occurred'
@@ -101,22 +73,10 @@ export default function CreationTagModalContent({
 							title: Liferay.Language.get('error'),
 							type: 'danger',
 						});
-					}
-				})
-				.catch(() => {
-					openToast({
-						message: Liferay.Language.get(
-							'an-unexpected-error-occurred'
-						),
-						title: Liferay.Language.get('error'),
-						type: 'danger',
 					});
-				});
-			resetForm();
-			setCheckbox(true);
-		},
-		validate: (values) => {
-			if (!checkbox) {
+				resetForm();
+			},
+			validate: (values) => {
 				validate(
 					{
 						assetLibraryIds: [required],
@@ -124,34 +84,8 @@ export default function CreationTagModalContent({
 					},
 					values
 				);
-			}
-		},
-	});
-
-	const isChecked = (itemValue: string) => {
-		return (values.assetLibraryIds as string[]).includes(itemValue);
-	};
-
-	const handleCheckboxChange = (itemValue: any) => {
-		setFieldValue(
-			'assetLibraryIds',
-			isChecked(itemValue)
-				? values.assetLibraryIds.filter((id) => id !== itemValue)
-				: [...values.assetLibraryIds, itemValue]
-		);
-	};
-
-	useEffect(() => {
-		if (checkbox) {
-			setFieldValue(
-				'assetLibraryIds',
-				allSpaces.flatMap((item) => item.value)
-			);
-		}
-		else {
-			setFieldValue('assetLibraryIds', []);
-		}
-	}, [checkbox, allSpaces, setFieldValue]);
+			},
+		});
 
 	return (
 		<form onSubmit={handleSubmit}>
@@ -169,81 +103,7 @@ export default function CreationTagModalContent({
 					value={values.tagName}
 				/>
 
-				<label htmlFor="multiSelect">
-					{Liferay.Language.get('space')}
-
-					<span className="ml-1 reference-mark">
-						<ClayIcon symbol="asterisk" />
-					</span>
-				</label>
-
-				{checkbox && (
-					<ClayMultiSelect
-						disabled={true}
-						id="multiSelect"
-						items={allSpaces}
-					/>
-				)}
-
-				{!checkbox && (
-					<ClayMultiSelect
-						disabled={checkbox}
-						id="multiSelect"
-						loadingState={3}
-						onChange={handleChange}
-						onItemsChange={(items: any) => {
-							setFieldValue(
-								'assetLibraryIds',
-								items.map((item: any) => item.value)
-							);
-						}}
-						sourceItems={availableSpaces}
-					>
-						{(item) => (
-							<ClayMultiSelect.Item
-								key={item.value}
-								textValue={item.label}
-							>
-								<div className="autofit-row autofit-row-center">
-									<div className="autofit-col">
-										<ClayCheckbox
-											aria-label={item.label}
-											checked={isChecked(item.value)}
-											onChange={() => {
-												handleCheckboxChange(
-													item.value
-												);
-											}}
-										/>
-									</div>
-
-									<div className="autofit-col">
-										<ClaySticker
-											displayType="outline-2"
-											size="sm"
-										>
-											{item.label.charAt(0).toUpperCase()}
-										</ClaySticker>
-									</div>
-
-									<div className="autofit-col">
-										{item.label}
-									</div>
-								</div>
-							</ClayMultiSelect.Item>
-						)}
-					</ClayMultiSelect>
-				)}
-
-				<div className="mt-2">
-					<ClayCheckbox
-						checked={checkbox}
-						label={Liferay.Language.get(
-							'make-this-tag-available-in-all-spaces'
-						)}
-						onChange={() => setCheckbox(!checkbox)}
-					/>
-				</div>
+				<CategorizationSpaces />
 			</ClayModal.Body>
 
 			<ClayModal.Footer
